added user roles and search feature

+++ routes/articles.js
@@ -1,5 +1,6 @@
 const express = require("express");
 const Article = require("./../models/Articles");
+const { ensureAdminAuthenticated } = require("../middleware/auth");
 const router = express.Router();
 
 router.get("/", async (req, res) => {
@@ -7,13 +8,23 @@ router.get("/", async (req, res) => {
   res.render("articles/index", { articles: articles });
 });
 
-router.get("/new", (req, res) => {
-  res.render("articles/new", { article: new Article() });
+router.get("/new", ensureAdminAuthenticated, (req, res) => {
+  if (req.user.isAdmin) {
+    res.render("articles/new", { article: new Article() });
+  } else {
+    req.flash("error_msg", "You are not admin and can't perform that action!");
+    res.redirect("/admin/login");
+  }
 });
 
-router.get("/edit/:id", async (req, res) => {
-  const article = await Article.findById(req.params.id);
-  res.render("articles/edit", { article: article });
+router.get("/edit/:id", ensureAdminAuthenticated, async (req, res) => {
+  if (req.user.isAdmin) {
+    const article = await Article.findById(req.params.id);
+    res.render("articles/edit", { article: article });
+  } else {
+    req.flash("error_msg", "You are not admin and can't perform that action!");
+    res.redirect("/admin/login");
+  }
 });
 
 router.get("/:slug", async (req, res) => {
@@ -40,9 +51,14 @@ router.put(
   saveArticleAndRedirect("edit")
 );
 
-router.delete("/:id", async (req, res) => {
-  await Article.findByIdAndDelete(req.params.id);
-  res.redirect("/");
+router.delete("/:id", ensureAdminAuthenticated, async (req, res) => {
+  if (req.user.isAdmin) {
+    await Article.findByIdAndDelete(req.params.id);
+    res.redirect("/");
+  } else {
+    req.flash("error_msg", "You are not admin and can't perform that action!");
+    res.redirect("/admin/login");
+  }
 });
 
 function saveArticleAndRedirect(path) {
@@ -51,6 +67,8 @@ function saveArticleAndRedirect(path) {
     article.title = req.body.title;
     article.description = req.body.description;
     article.markdown = req.body.markdown;
+    (article.author.id = req.user._id),
+      (article.author.username = req.user.username);
     try {
       article = await article.save();
       res.redirect(`/articles/${article.slug}`);
